// apps/web/src/api/generated/model/taskResponseDtoStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * BTN HRMS API
 * The Human Resource Management System API description
 * OpenAPI spec version: 1.0
 */

export type TaskResponseDtoStatus = typeof TaskResponseDtoStatus[keyof typeof TaskResponseDtoStatus];


export const TaskResponseDtoStatus = {
  created: 'created',
  assigned: 'assigned',
  in_progress: 'in_progress',
  declined: 'declined',
  submitted: 'submitted',
  revision: 'revision',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;